Minor Ui fixes
