+++ src/components/Products/index.jsx
@@ -1,0 +1,14 @@
+import "../../styles/Products.css";
+import ProductsCard from "../Cards/ProductsCard";
+
+const Products = () => {
+  return (
+    <div className="group">
+      <h1>PRODUCTS</h1>
+      <hr />
+      <ProductsCard />
+    </div>
+  );
+};
+
+export default Products;
