+++ src/components/SingleClub/index.jsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import ClubUpperImage from "../assets/pictures/ClubUpperImage.svg";
+import ClubUpperImage from "../../assets/pictures/ClubUpperImage.svg";
 import { useLocation } from "react-router-dom";
 import "react-toastify/dist/ReactToastify.css";
-import displayRazorpay from "../service/PaymentGateway";
+import displayRazorpay from "../../service/PaymentGateway";
 
 export default function SingleClub({ club }) {
   const [money, setmoney] = React.useState({ donatedmoney: 0 });
