+++ src/components/index.jsx
@@ -1,0 +1,14 @@
+export { default as Banner } from "./Banners/Banner";
+export { default as ClubBanner } from "./Banners/ClubBanner";
+export { default as DonateBanner } from "./Banners/DonateBanner";
+export { default as EventsBanner } from "./Events";
+export { default as Footer } from "./Footer";
+export { default as HomeCardsContainer } from "./HomeCardsContainer";
+export { default as LoginBanner } from "./Banners/LoginBanner";
+export { default as Navbar } from "./Navbar";
+export { default as Products } from "./Products";
+export { default as ProductsCard } from "./Cards/ProductsCard";
+export { default as ShopHeader } from "./Headers/ShopHeader";
+export { default as SingleClub } from "./SingleClub";
+export { default as SingleEvent } from "./Events/SingleEvent";
+export { default as SocialIcons } from "./SocialIcons";
