+++ src/pages/clubs/ClubsRegister.jsx
@@ -32,7 +32,7 @@ const ClubLogin = () => {
     address: "",
     pincode: "",
     description: "",
-    tagLine: ""
+    tagLine: "",
   });
 
   const [isEmailValid, setIsEmailValid] = useState(false);
@@ -52,7 +52,7 @@ const ClubLogin = () => {
     e.preventDefault();
 
     setIsLoading(true);
-    await RegisterClub({ ...credentials})
+    await RegisterClub({ ...credentials });
     setIsLoading(false);
 
     toast("🌈 Registered club !", {
@@ -143,7 +143,6 @@ const ClubLogin = () => {
                   value={credentials.tagLine}
                   onChange={handleChange}
                   required
-                  autoFocus
                   aria-label="Club Tagline"
                   id="club-tagLine"
                   maxLength={50}
@@ -175,7 +174,6 @@ const ClubLogin = () => {
                   value={credentials.email}
                   onChange={handleChange}
                   required
-                  autoFocus
                   aria-label="Club email"
                   id="email-des"
                 />
@@ -212,7 +210,6 @@ const ClubLogin = () => {
                     value={credentials.password}
                     onChange={handleChange}
                     required
-                    autoFocus
                     id="password-des"
                   />
                   <input
@@ -241,7 +238,6 @@ const ClubLogin = () => {
                     value={credentials.confirmPassword}
                     onChange={handleChange}
                     required
-                    autoFocus
                     id="confirm-password-des"
                   />
