+++ src/pages/display/EventsPage.jsx
@@ -1,11 +1,9 @@
 import React, { useEffect, useState } from "react";
 import "../../styles/ClubsPage.css";
-import Navbar from "../../components/Navbar";
-import Footer from "../../components/Footer";
+import { Navbar, Footer, SingleEvent } from "../../components";
 import { GetAllEvents } from "../../service/MilanApi";
-import SingleEvent from "../../components/SingleEvent";
 import Eventspic from "../../assets/pictures/EventsPagefloating.svg";
-import '../../styles/EventsPage.css';
+import "../../styles/EventsPage.css";
 
 const EventsPage = () => {
   document.title = "Milan | Events";
@@ -24,11 +22,12 @@ const EventsPage = () => {
     <>
       <Navbar />
 
-
-
       {/* Banner Portion Code  */}
 
-      <div id="event_banner" className="container d-flex flex-column flex-lg-row text-center text-lg-start">
+      <div
+        id="event_banner"
+        className="container d-flex flex-column flex-lg-row text-center text-lg-start"
+      >
         <div id="eventCol_2" className="d-none d-sm-block">
           <img src={Eventspic} alt="" className="event_image" />
         </div>
