+++ src/pages/index.jsx
@@ -1,0 +1,14 @@
+export { default as AboutUs } from "./AboutUs";
+export { default as ContactPage } from "./ContactUs";
+export { default as Home } from "./Home";
+export { default as Donate } from "./Donate";
+export { default as UserLogin } from "./user/UserLogin";
+export { default as UserProfile } from "./user/UserProfile";
+export { default as UserRegister } from "./user/UserRegister";
+export { default as ShopLanding } from "./shops/ShopLanding";
+export { default as ClubsPage } from "./display/ClubsPage";
+export { default as EventsPage } from "./display/EventsPage";
+export { default as ClubRegister } from "./clubs/ClubsRegister";
+export { default as EventCreate } from "./clubs/EventCreate";
+export { default as ClubLogin } from "./clubs/ClubsLogin";
+export { default as ClubProfile } from "./clubs/ClubProfile";
