+++ src/pages/shops/ShopLanding.jsx
@@ -1,20 +1,13 @@
-import React from "react";
-import ShopHeader from "../../components/ShopHeader";
-import Navbar from "../../components/Navbar";
-import Products from "../../components/Products";
-
-
+import { Navbar, Products, ShopHeader } from "../../components";
 
 const ShopLanding = () => {
+  return (
+    <>
+      <Navbar />
+      <ShopHeader />
+      <Products />
+    </>
+  );
+};
 
-
-    return (
-        <>
-          <Navbar />
-          <ShopHeader /> 
-          <Products /> 
-        </>
-    )
-}
-
-export default ShopLanding
+export default ShopLanding;
