+++ src/styles/ClubsPage.css
@@ -96,7 +96,7 @@ div#clubs_banner {
     max-width: 270px !important;
     width: 80%;
   }
-  
+
   #clubscol_1 p {
     margin-inline: 20px;
   }
@@ -179,23 +179,8 @@ div#clubs_banner {
   }
 }
 
-body::-webkit-scrollbar-track {
-  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  border-radius: 10px;
-  background-color: #f5f5f5;
-}
-
 body::-webkit-scrollbar {
-  width: 12px;
-  background-color: #f5f5f5;
-}
-
-body::-webkit-scrollbar-thumb {
-  border-radius: 10px;
-  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  background-color: darkgrey;
+  display: none;
 }
 
 .eventspageintro_parent {
