+++ src/styles/DonateBanner.css
@@ -2,7 +2,7 @@
 
 div#donate-banner {
   width: 80%;
-  margin: 64px auto;
+  margin: 60px auto;
   font-family: "Mulish", sans-serif;
 }
 
@@ -10,37 +10,30 @@ div#donateCol1 h1 {
   font-style: normal;
   font-weight: bold;
   font-size: 48px;
-  line-height: 60px;
-
   color: #297f75;
 }
 
 div#donateCol1 p {
   font-style: normal;
-  font-weight: bold;
-  font-size: 20px;
-  line-height: 30px;
-
+  font-weight: normal;
+  font-size: 18px;
   color: #000000;
+  max-width: 65%;
 }
 
 div#donateCol1 button {
-  width: 196px;
-  padding: 8px;
+  padding: 8px 12px;
   border: none;
   border-radius: 12px;
-  font-style: normal;
-  font-weight: 800;
+  font-weight: 600;
   font-size: 18px;
-  line-height: 30px;
-  text-align: center;
   color: #ffffff;
   background-color: #297f75;
-  transition: 0.2s;
+  transition: background-color 0.5 ease-in-out;
 }
 
 div#donateCol1 button:hover {
-  background-color: #0a3f39;
+  background-color: #22756b;
 }
 
 .button_animation:hover {
